move the writing of regional boundaries to the west-northeast-southeast division
